unittest + restructing
